requirements.txt
torch
torchvision
torchaudio
gradio==3.50.0
numpy==1.26.4
opencv-python
matplotlib
scikit-image
moviepy
Pillow
transformers
addict
yapf
timm==0.4.5
supervision==0.6.0
pycocotools
groundingdino-py==0.4.0
gdown
IPython
wget
psutil>=5.9.0
packaging>=21.0 